Make monasca-notification Py35 compatible
Adjust the monasca-notification to run
under Python3.5

Story: 2000975
Task: 4130

+++ monasca_notification/notification.py
@@ -65,7 +65,7 @@ class Notification(object):
         self.alarm_id = alarm['alarmId']
         self.alarm_name = alarm['alarmName']
         # The event timestamp is in milliseconds
-        self.alarm_timestamp = alarm['timestamp'] / 1000
+        self.alarm_timestamp = int(alarm['timestamp'] / 1000.0)
         self.message = alarm['stateChangeReason']
         self.state = alarm['newState']
         self.severity = alarm['severity']
